separated edit medication dialog to separate file
